pyproject.toml
[tool.poetry]
name = "telegram-food-parser"
version = "0.1.0"
description = "Парсер food-каналов Telegram"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
telethon = ">=1.28.5"
aiogram = ">=3.0,<4.0"
python-dotenv = ">=1.0,<2.0"
beautifulsoup4 = ">=4.12.2"
transformers = ">=4.30.0"
torch = ">=2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
ruff = "^0.0.291"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"] 